Centralise disaster and disaster event models
