+++ packages/database/src/modelClasses/DisasterEvent.js
@@ -1,9 +1,11 @@
 /**
- * Tupaia MediTrak
- * Copyright (c) 2019 Beyond Essential Systems Pty Ltd
- **/
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
 
-import { DatabaseType, TYPES, DatabaseModel } from '@tupaia/database';
+import { DatabaseModel } from '../DatabaseModel';
+import { DatabaseType } from '../DatabaseType';
+import { TYPES } from '../types';
 
 class DisasterEventType extends DatabaseType {
   static databaseType = TYPES.DISASTER_EVENT;
