--- packages/web-config-server/src/models/Disaster.js
@@ -1,26 +1,0 @@
-/**
- * Tupaia Config Server
- * Copyright (c) 2018 Beyond Essential Systems Pty Ltd
- **/
-
-import { TYPES } from '@tupaia/database';
-import { BaseModel } from './BaseModel';
-
-export class Disaster extends BaseModel {
-  static databaseType = TYPES.DISASTER;
-
-  static fields = ['id', 'type', 'description', 'name', 'countryCode'];
-
-  static getDisasters() {
-    return Disaster.database.executeSql(`
-      SELECT
-        *,
-        "disaster".type,
-        ST_AsGeoJSON("point") as point,
-        ST_AsGeoJSON("bounds") as bounds
-      FROM "disaster"
-        LEFT JOIN "entity"
-          ON "disaster"."id" = "entity"."code";
-    `);
-  }
-}
